Rename regression scripts.
The regression scripts are bash-python hybrid scripts.  The .sh
extension was casing some validators to validate it as a shell script
but raise an error when it encountered the python code portion of the
script.  Renaming the script to .py should validate the script as a
python script and validate it as a valid python script since the shell
script portion is embedded as string within valid python code.